+++ CHANGELOG.md
@@ -7,7 +7,7 @@ From version 2.6.0, the sections in this file adhere to the [keep a changelog](h
 * [#1937](https://github.com/Shopify/shopify-cli/pull/1937): Fix `theme pull` to no longer add empty lines on Windows
 
 ### Added
-* [#1892](https://github.com/Shopify/shopify-cli/pull/1892): Add `-F`/`--filter` parameter to filter files on `theme push`/`theme pull` commands
+* [#1892](https://github.com/Shopify/shopify-cli/pull/1892): Add `-o`/`--only` parameter to filter files on `theme push`/`theme pull` commands
 
 ## 2.9.0